Move the OpenBSD-derived base64.h file into the src/ folder.
