+++ src/openbsd-compat/base64.h
